plot holes
After 2 days of debugging our segment allocator I was trying
to optimize the allocator so that if we run out of memory we can
fallback on regular malloc, however, to check if we are out or not
it required to transverse every single block and took up so much time
that even our fps was dropping hard, to make things easier I've
decided to move the parser into a separate thread and to be
processed only on demand. This resolved the issue and this thread will
keep the rover pointing at the biggest free block such that we waste
mininum memory too. Now there is almost to 0 cost of doing a malloc fallback
and essentially has an overhead of 1 if statement. The new thread has yet
to be optimized, as right now the algorithms it uses is brute force.